Revert "..."
This reverts commit fdae8e35f31ba20ebad37d79224444a843b707ae.

+++ src/FrameworkResources.cpp
@@ -118,7 +118,7 @@ static Sexy::ResoddedFrameworkResourceId ResoddedFrameworkGetIdByVariable(const 
 			aMap[*(int*)gResources[i]] = i;
 	}
 
-	MyMap::iterator anItr = aMap.find((int)(uintptr_t)theVariable);
+	MyMap::iterator anItr = aMap.find((int)theVariable);
 	if (anItr == aMap.end())
 		return ResoddedFrameworkRESOURCE_ID_MAX;
 	else
