+++ src/Resources.cpp
@@ -3308,7 +3308,7 @@ static Sexy::ResourceId GetIdByVariable(const void *theVariable)
 			aMap[*(int *)gResources[i]] = i;
 	}
 
-	MyMap::iterator anItr = aMap.find((int)(uintptr_t)theVariable);
+	MyMap::iterator anItr = aMap.find((int)theVariable);
 	if (anItr == aMap.end())
 		return RESOURCE_ID_MAX;
 	else
